requirements.txt
alembic==1.12.1
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
exceptiongroup==1.1.3
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
Werkzeug==3.0.1
